Fix linting errors
For now, some of the errors are explicitly ignored, always per line.

For the explicit-any violations: those will need to be fixed

For the unused-var violations: those will probably stay, either per line
or per file, as they are only in test files and aren't so much variables
as well function signatures for mock. Need to figure out how to inform
eslint of that...